api certification

+++ src/main/java/com/web/bookstore/api/dto/BookSummary.java
@@ -1,7 +1,9 @@
 package com.web.bookstore.api.dto;
 
+import java.math.BigDecimal;
+
 public record BookSummary(
         Long id,
         String title,
-        Integer price
-) { }
+        BigDecimal price
+) {}
